+++ Content.Goobstation.Server/ServerCurrency/ServerCurrencySystem.Inky.cs
@@ -2,23 +2,35 @@ using Content.Server.GameTicking;
 using Content.Shared.Humanoid;
 using Content.Shared.Mind;
 using Content.Shared.Mind.Components;
+using Content.Shared.Players;
+using Content.Shared.Roles.Jobs;
 using Content.Shared.Silicons.Borgs.Components;
+using Robust.Server.Player;
+using Robust.Shared.Player;
 
 namespace Content.Goobstation.Server.ServerCurrency;
 
 public sealed partial class ServerCurrencySystem
 {
+    [Dependency] private SharedMindSystem _mind = default!;
+    [Dependency] private SharedJobSystem _jobs = default!;
+    [Dependency] private IPlayerManager _players = default!;
+    [Dependency] private SharedPlayerSystem _playerSystem = default!;
+
+    private int _goobcoinsNonAntagMultiplier = 1;
+    private int _goobcoinsServerMultiplier = 1;
+    private int _goobcoinsMinPlayers = 0;
+
+
     public void InitializeInky()
     {
-        SubscribeLocalEvent<RoundEndTextAppendEvent>(OnRoundEndText);
+        SubscribeLocalEvent<RoundEndTextAppendEvent>(OnRoundEndTextAppend);
     }
 
-    private void OnRoundEndText(RoundEndTextAppendEvent ev)
+    private void OnRoundEndTextAppend(RoundEndTextAppendEvent ev)
         {
             if (_players.PlayerCount < _goobcoinsMinPlayers) // 0 default unless changed
                 return;
-
-            // var lowPopMultiplier = 1.0 - (_players.PlayerCount / (double)_players.MaxPlayers); // no lowpop multiplier for you chuddie
 
             var query = EntityQueryEnumerator<MindContainerComponent>();
 
@@ -38,10 +50,10 @@ public sealed partial class ServerCurrencySystem
                         && mind.OriginalOwnerUserId.HasValue
                         && _players.TryGetSessionById(mind.UserId, out var session))
                     {
-                        int money = _goobcoinsPerPlayer;
+                        int money = 0;
                         if (session is not null)
                         {
-                            money += _jobs.GetJobGoobcoins(session);
+                            money += GetJobGoobcoins(session);
                             if (!_jobs.CanBeAntag(session))
                                 money *= _goobcoinsNonAntagMultiplier;
                         }
@@ -57,4 +69,13 @@ public sealed partial class ServerCurrencySystem
                 }
             }
         }
+
+    public int GetJobGoobcoins(ICommonSession player)
+    {
+        if (_playerSystem.ContentData(player) is not { Mind: { } mindId }
+            || !_jobs.MindTryGetJob(mindId, out var prototype))
+            return 1;
+
+        return prototype.Currency;
+    }
 }
