+++ Content.Trauma.Client/RoundEndCredits/EndRoundCreditsControl.xaml.cs
@@ -1,5 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-
+// <inky>
+using Content.Shared.Dataset;
+using Robust.Shared.Random;
+// </inky>
 using System.Linq;
 using Content.Client.GameTicking.Managers;
 using Content.Shared.GameTicking;
@@ -10,10 +13,6 @@ using Robust.Client.ResourceManagement;
 using Robust.Client.UserInterface.Controls;
 using Robust.Client.UserInterface.XAML;
 using Robust.Shared.Utility;
-// inky
-using Content.Shared.Dataset;
-using Robust.Shared.Random;
-// /inky
 
 namespace Content.Trauma.Client.RoundEndCredits;
 
@@ -22,9 +21,9 @@ public sealed partial class EndRoundCreditsControl : ScrollContainer
 {
     [Dependency] private IEntitySystemManager _entitySystem = default!;
     private readonly ClientGameTicker _gameTicker;
-    // inky
+    // <inky>
     [Dependency] private IRobustRandom _gamba = default!;
-    // /iniky
+    // </inky>
 
     private static readonly ResPath Logo = new("/Textures/Logo/logo.png");
     private static readonly ResPath Pixellari = new("/Fonts/_Trauma/Pixellari.ttf");
@@ -35,9 +34,9 @@ public sealed partial class EndRoundCreditsControl : ScrollContainer
     private const int BigFontSize = 24;
     private const int HeaderFontSize = 42;
 
-    // inky
-    private const string datasetId = "EndRoundCredits";
-    // /inky
+    // <inky>
+    private ProtoId<LocalizedDatasetPrototype> datasetId = "EndRoundCredits";
+    // </inky>
 
     public EndRoundCreditsControl()
     {
@@ -80,11 +79,11 @@ public sealed partial class EndRoundCreditsControl : ScrollContainer
         // Image mgsv episode number and jargon
         ServerImageBox.AddChild(serverImage);
         EpisodeNumber.Text = Loc.GetString("round-end-credits-trauma-episode", ("roundid", message.RoundId), ("title", message.GamemodeTitle));
-        // inky start
+        // <Inky>
         // IntroJargonLabel.Text = Loc.GetString("round-end-credits-trauma-jargon", ("station", stationName));
-        if (proto.TryIndex<LocalizedDatasetPrototype>(datasetId, out var datasetPrototype))
-            IntroJargonLabel.Text = Loc.GetString(_gamba.Pick(datasetPrototype.Values));
-        // /inky
+        if (proto.TryIndex(datasetId, out var datasetPrototype))
+            IntroJargonLabel.Text = _gamba.Pick(proto.Index(datasetId).Values);
+        // </Inky>
         // The fonts
         EpisodeNumber.FontOverride = bigFont;
         IntroJargonLabel.FontOverride = bigFont;
